requirements.txt
streamlit
elasticsearch
pandas
python-dotenv